+++ src/main/java/org/drugis/addis/trialverse/service/TrialverseDataService.java
@@ -6,7 +6,7 @@ import org.drugis.addis.trialverse.model.TrialData;
  * Created by connor on 9-5-14.
  */
 public interface TrialverseDataService {
-  TrialData getTrialData(Integer namespaceId, String outcomeUri);
+  TrialData getTrialData(Long namespaceId, String outcomeUri);
 
-  TrialData getTrialData(Integer namespaceId);
+  TrialData getTrialData(Long namespaceId);
 }
